refactored subset process
